+++ src/mael/teensy/serial.cpp
@@ -51,7 +51,7 @@ static bool ExecuteCommand(char *cmd, char *arg0, char *arg1, char *arg2, char *
             return false;
         if (!ParseLong(arg2, 0, 1000, &w))
 
-        SetDriveSpeed((float)x, (float)y, (float)w);
+        SetDriveSpeed((double)x, (double)y, (double)w);
         return true;
     } else {
         Serial.println("Invalid command or arguments");
